test(ssr): refactor ssr render tests
Merge renderToString and renderToStream to run the same tests
